requirements.txt
django-material == 1.12.0
Pillow == 10.0.1
django-imagekit == 5.0.0
